Move ElasticInterval object in separated file, fix some problems.
